requirements.txt
alembic==1.4.2
appdirs==1.4.4
asttokens==2.0.4
attrs==19.3.0
Authlib==0.14.3
black==19.10b0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
executing==0.5.4
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
icecream==2.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.0
pathspec==0.8.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
PyMySQL==0.10.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.1
